:lipstick: Add Signature Verification Rendering to Decoded Token Table
As part of this, DecodedToken.verified has been changed so that it has
three states. If there is a public key during decoding then this value
will either be True or False, otherwise it will be None which indicates
that signature verification has not been performed. Unit Tests have been
updated to reflect these changes by switching from assertFalse to assetIs
methods.